requirements.txt
numpy==1.26.4
tensorflow==2.17.1
tqdm==4.67.1
pillow==11.0.0
transformers==4.48.3
torch==2.6.0
streamlit==1.40.1